refactor model name
